requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
distro==1.9.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-core==0.3.74
langchain-mcp-adapters==0.1.9
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langgraph==0.6.5
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.0
langsmith==0.4.14
markdown==3.8.2
markdown-it-py==4.0.0
mcp==1.13.0
mcp-server-tapd==8.0.60
mdurl==0.1.2
openai==1.99.9
orjson==3.11.2
ormsgpack==1.10.0
packaging==25.0
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
pyyaml==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.0
shellingham==1.5.4
sniffio==1.3.1
sqlalchemy==2.0.43
sse-starlette==3.0.2
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
xxhash==3.5.0
zstandard==0.23.0
